requirements.txt
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
ta-lib==0.4.28
matplotlib==3.7.2
ccxt==4.1.56
websockets==12.0
python-dotenv==1.0.0